PEP 8 compliance
Made modifications to the files to make them compliant with Python's PEP 8:

- http://www.python.org/dev/peps/pep-0008/

As part of this some of the functions were renamed to use a lowercase-
underscore format instead of camelCase. The camelCase functions were
simply made deprecated wrappers for the real functions. As a note, the wrapper
functions do throw a 'DeprecationWarning' when called.

The docs were also updated to the new function names as well.